refactor: rename all StudentEventsList files to EventList
