+++ src/components/StudentSection/index.ts
@@ -1,3 +1,3 @@
 export { Reward } from './Reward/Reward'
 export { StudentEventDetail } from './StudentEventDetail/StudentEventDetail'
-export { StudentEventList } from './EventList/StudentEventsList'
+export { EventList } from './EventList/EventList'
