+++ src/components/index.ts
@@ -8,7 +8,7 @@ export { ContactsMap } from './ContactsMap/ContactsMap'
 export { CopyableText } from './CopyableText/CopyableText'
 export { Divider } from './Divider/Divider'
 export { EventForm, EventDetail } from './Events'
-export { StudentEventList, StudentEventDetail, Reward } from './StudentSection'
+export { EventList, StudentEventDetail, Reward } from './StudentSection'
 export { FavoritesList } from './FavoritesList/FavoritesList'
 export { HomeLayout } from './HomeLayout/HomeLayout'
 export { IconButton } from './IconButton/IconButton'
